requirements.txt
    streamlit
    langchain
    langchain-core
    langchain-community
    langchain-huggingface
    langchain-google-genai
    langchain-chroma
    chromadb
    python-dotenv
    gitpython
    sentence-transformers
    firebase-admin
    google-cloud-firestore
